+++ composeunstyled-theming/src/commonMain/kotlin/com/composeunstyled/theme/ColoredIndication.kt
@@ -1,7 +1,14 @@
 package com.composeunstyled.theme
 
+import androidx.compose.animation.core.Animatable
+import androidx.compose.animation.core.AnimationSpec
+import androidx.compose.animation.core.snap
 import androidx.compose.foundation.IndicationNodeFactory
-import androidx.compose.foundation.interaction.*
+import androidx.compose.foundation.interaction.DragInteraction
+import androidx.compose.foundation.interaction.FocusInteraction
+import androidx.compose.foundation.interaction.HoverInteraction
+import androidx.compose.foundation.interaction.InteractionSource
+import androidx.compose.foundation.interaction.PressInteraction
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.remember
 import androidx.compose.ui.Modifier
@@ -9,8 +16,9 @@ import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.drawscope.ContentDrawScope
 import androidx.compose.ui.node.DelegatableNode
 import androidx.compose.ui.node.DrawModifierNode
-import androidx.compose.ui.node.invalidateDraw
+import kotlinx.coroutines.delay
 import kotlinx.coroutines.launch
+import kotlin.time.TimeSource
 
 @Composable
 fun rememberColoredIndication(
@@ -18,13 +26,17 @@ fun rememberColoredIndication(
     pressedColor: Color = Color.Unspecified,
     focusedColor: Color = Color.Unspecified,
     draggedColor: Color = Color.Unspecified,
+    showAnimationSpec: AnimationSpec<Float> = snap(),
+    hideAnimationSpec: AnimationSpec<Float> = snap()
 ): IndicationNodeFactory {
     return remember {
         ColoredIndication(
             hoveredColor = hoveredColor,
             pressedColor = pressedColor,
             focusedColor = focusedColor,
-            draggedColor = draggedColor
+            draggedColor = draggedColor,
+            animationSpecEnter = showAnimationSpec,
+            animationSpecExit = hideAnimationSpec
         )
     }
 }
@@ -36,13 +48,17 @@ fun rememberColoredIndication(
     focusedAlpha: Float = 0.1f,
     hoveredAlpha: Float = 0.08f,
     pressedAlpha: Float = 0.1f,
+    showAnimationSpec: AnimationSpec<Float> = snap(),
+    hideAnimationSpec: AnimationSpec<Float> = snap()
 ): IndicationNodeFactory {
     return remember {
         ColoredIndication(
             hoveredColor = color.copy(alpha = hoveredAlpha),
             pressedColor = color.copy(alpha = pressedAlpha),
             focusedColor = color.copy(alpha = focusedAlpha),
-            draggedColor = color.copy(alpha = draggedAlpha)
+            draggedColor = color.copy(alpha = draggedAlpha),
+            animationSpecEnter = showAnimationSpec,
+            animationSpecExit = hideAnimationSpec
         )
     }
 }
@@ -52,10 +68,19 @@ class ColoredIndication(
     private val pressedColor: Color,
     private val focusedColor: Color,
     private val draggedColor: Color,
+    private val animationSpecEnter: AnimationSpec<Float>,
+    private val animationSpecExit: AnimationSpec<Float>
 ) : IndicationNodeFactory {
 
-    override fun create(interactionSource: InteractionSource): DelegatableNode =
-        ColoredIndicationInstance(interactionSource, hoveredColor, pressedColor, focusedColor, draggedColor)
+    override fun create(interactionSource: InteractionSource): DelegatableNode = ColoredIndicationInstance(
+        interactionSource = interactionSource,
+        hoveredColor = hoveredColor,
+        pressedColor = pressedColor,
+        focusedColor = focusedColor,
+        draggedColor = draggedColor,
+        animationSpecEnter = animationSpecEnter,
+        animationSpecExit = animationSpecExit
+    )
 
     override fun hashCode(): Int = -1
 
@@ -67,72 +92,159 @@ class ColoredIndication(
         private val pressedColor: Color,
         private val focusedColor: Color,
         private val draggedColor: Color,
-    ) :
-        Modifier.Node(), DrawModifierNode {
-        private var isPressed = false
-        private var isHovered = false
-        private var isFocused = false
-        private var isDragged = false
+        private val animationSpecEnter: AnimationSpec<Float>,
+        private val animationSpecExit: AnimationSpec<Float>
+    ) : Modifier.Node(), DrawModifierNode {
+
+        private val pressedAlpha = Animatable(0f)
+        private val hoveredAlpha = Animatable(0f)
+        private val focusedAlpha = Animatable(0f)
+        private val draggedAlpha = Animatable(0f)
+
+        private val timeSource = TimeSource.Monotonic
+        private var pressStartTime: TimeSource.Monotonic.ValueTimeMark? = null
+        private var hoverStartTime: TimeSource.Monotonic.ValueTimeMark? = null
+        private var focusStartTime: TimeSource.Monotonic.ValueTimeMark? = null
+        private var dragStartTime: TimeSource.Monotonic.ValueTimeMark? = null
+        private val minimumVisualDuration = 25L
 
         override fun onAttach() {
             coroutineScope.launch {
-                var pressCount = 0
-                var hoverCount = 0
-                var focusCount = 0
-                var dragCount = 0
                 interactionSource.interactions.collect { interaction ->
                     when (interaction) {
-                        is PressInteraction.Press -> pressCount++
-                        is PressInteraction.Release -> pressCount--
-                        is PressInteraction.Cancel -> pressCount--
-                        is HoverInteraction.Enter -> hoverCount++
-                        is HoverInteraction.Exit -> hoverCount--
-                        is FocusInteraction.Focus -> focusCount++
-                        is FocusInteraction.Unfocus -> focusCount--
-                        is DragInteraction.Start -> dragCount++
-                        is DragInteraction.Stop -> dragCount--
-                    }
-                    val pressed = pressCount > 0
-                    val hovered = hoverCount > 0
-                    val focused = focusCount > 0
-                    val dragged = focusCount > 0
+                        // press
+                        is PressInteraction.Press -> {
+                            launch {
+                                pressStartTime = timeSource.markNow()
+                                pressedAlpha.animateTo(1f, animationSpecEnter)
+                            }
+                        }
 
-                    var invalidateNeeded = false
-                    if (isPressed != pressed) {
-                        isPressed = pressed
-                        invalidateNeeded = true
+                        is PressInteraction.Release -> {
+                            launch {
+                                val elapsed = pressStartTime?.elapsedNow()?.inWholeMilliseconds ?: 0L
+                                val remaining = minimumVisualDuration - elapsed
+
+                                if (remaining > 0) {
+                                    delay(remaining)
+                                }
+
+                                pressedAlpha.animateTo(0f, animationSpecExit)
+                                pressStartTime = null
+                            }
+                        }
+
+                        is PressInteraction.Cancel -> {
+                            launch {
+                                val elapsed = pressStartTime?.elapsedNow()?.inWholeMilliseconds ?: 0L
+                                val remaining = minimumVisualDuration - elapsed
+
+                                if (remaining > 0) {
+                                    delay(remaining)
+                                }
+
+                                pressedAlpha.animateTo(0f, animationSpecExit)
+                                pressStartTime = null
+                            }
+                        }
+
+                        // hover
+                        is HoverInteraction.Enter -> {
+                            launch {
+                                hoverStartTime = timeSource.markNow()
+                                hoveredAlpha.animateTo(1f, animationSpecEnter)
+                            }
+                        }
+
+                        is HoverInteraction.Exit -> {
+                            launch {
+                                val elapsed = hoverStartTime?.elapsedNow()?.inWholeMilliseconds ?: 0L
+                                val remaining = minimumVisualDuration - elapsed
+
+                                if (remaining > 0) {
+                                    delay(remaining)
+                                }
+
+                                hoveredAlpha.animateTo(0f, animationSpecExit)
+                                hoverStartTime = null
+                            }
+                        }
+
+                        // focus
+                        is FocusInteraction.Focus -> {
+                            launch {
+                                focusStartTime = timeSource.markNow()
+                                focusedAlpha.animateTo(1f, animationSpecEnter)
+                            }
+                        }
+
+                        is FocusInteraction.Unfocus -> {
+                            launch {
+                                val elapsed = focusStartTime?.elapsedNow()?.inWholeMilliseconds ?: 0L
+                                val remaining = minimumVisualDuration - elapsed
+
+                                if (remaining > 0) {
+                                    delay(remaining)
+                                }
+
+                                focusedAlpha.animateTo(0f, animationSpecExit)
+                                focusStartTime = null
+                            }
+                        }
+
+                        // drag
+                        is DragInteraction.Start -> {
+                            launch {
+                                dragStartTime = timeSource.markNow()
+                                draggedAlpha.animateTo(1f, animationSpecEnter)
+                            }
+                        }
+
+                        is DragInteraction.Stop -> {
+                            launch {
+                                val elapsed = dragStartTime?.elapsedNow()?.inWholeMilliseconds ?: 0L
+                                val remaining = minimumVisualDuration - elapsed
+
+                                if (remaining > 0) {
+                                    delay(remaining)
+                                }
+
+                                draggedAlpha.animateTo(0f, animationSpecExit)
+                                dragStartTime = null
+                            }
+                        }
+
+                        is DragInteraction.Cancel -> {
+                            launch {
+                                val elapsed = dragStartTime?.elapsedNow()?.inWholeMilliseconds ?: 0L
+                                val remaining = minimumVisualDuration - elapsed
+
+                                if (remaining > 0) {
+                                    delay(remaining)
+                                }
+
+                                draggedAlpha.animateTo(0f, animationSpecExit)
+                                dragStartTime = null
+                            }
+                        }
                     }
-                    if (isHovered != hovered) {
-                        isHovered = hovered
-                        invalidateNeeded = true
-                    }
-                    if (isFocused != focused) {
-                        isFocused = focused
-                        invalidateNeeded = true
-                    }
-                    if (isDragged != dragged) {
-                        isDragged = dragged
-                        invalidateNeeded = true
-                    }
-                    if (invalidateNeeded) invalidateDraw()
                 }
             }
         }
 
-
         override fun ContentDrawScope.draw() {
             drawContent()
-            if (isPressed) {
-                drawRect(color = pressedColor, size = size)
+            if (pressedAlpha.value > 0f) {
+                drawRect(color = pressedColor, alpha = pressedAlpha.value, size = size)
             }
-            if (isDragged) {
-                drawRect(color = draggedColor, size = size)
+            if (draggedAlpha.value > 0f) {
+                drawRect(color = draggedColor, alpha = draggedAlpha.value, size = size)
             }
-            if (isHovered) {
-                drawRect(color = hoveredColor, size = size)
+            if (hoveredAlpha.value > 0f) {
+                drawRect(color = hoveredColor, alpha = hoveredAlpha.value, size = size)
             }
-            if (isFocused) {
-                drawRect(color = focusedColor, size = size)
+            if (focusedAlpha.value > 0f) {
+                drawRect(color = focusedColor, alpha = focusedAlpha.value, size = size)
             }
         }
     }